tinyfed_sim/app/requirements.txt
paho-mqtt==2.1.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
psutil==5.9.8
matplotlib==3.8.4
